260120 translate/requirements.txt

streamlit>=1.28.0
transformers>=4.45.0
torch>=2.0.0
pdf2image>=1.16.3
Pillow>=10.0.0
pymupdf>=1.23.0
streamlit-option-menu
pandas
accelerate
sentencepiece
protobuf
